update delivery programme, programme manager list, catalog entity displayName
